# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pharmacy Manager API",
        "description": "A comprehensive API for managing pharmacy operations including products, categories, orders, customers, and authentication.",
        "contact": {
            "name": "Pharmacy Manager Support",
            "email": "admin@pharmacy.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get all products",
                "description": "Retrieve all products with pagination, search, and filtering options",
                "operationId": "getProducts",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term for product name, description, or SKU",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Filter by category ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "include",
                        "in": "query",
                        "description": "Include relationships (e.g., category)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "category"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Products retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Category": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Pain Relief"
                    },
                    "description": {
                        "type": "string",
                        "example": "Medications for pain management"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Product": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Category"
                    },
                    {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "Paracetamol 500mg"
                            },
                            "description": {
                                "type": "string",
                                "example": "Pain relief medication"
                            },
                            "price": {
                                "type": "number",
                                "format": "float",
                                "example": 5.99
                            },
                            "quantity": {
                                "type": "integer",
                                "example": 100
                            },
                            "sku": {
                                "type": "string",
                                "example": "PROD-001"
                            },
                            "manufacture_date": {
                                "type": "string",
                                "format": "date",
                                "example": "2024-01-15"
                            },
                            "expiry_date": {
                                "type": "string",
                                "format": "date",
                                "example": "2026-01-15"
                            },
                            "category_id": {
                                "type": "integer",
                                "example": 1
                            },
                            "created_at": {
                                "type": "string",
                                "format": "date-time"
                            },
                            "updated_at": {
                                "type": "string",
                                "format": "date-time"
                            },
                            "category": {
                                "$ref": "#/components/schemas/Category"
                            }
                        },
                        "type": "object"
                    }
                ]
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}